Homepage html bearbeitet

+++ src/main/java/com/example/springboot/Product.java
@@ -25,7 +25,7 @@ public class Product {
     private String name;
 
     @Nullable
-    private int price;
+    private double price;
 
     @Nullable
     private String owner;
@@ -34,7 +34,7 @@ public class Product {
 
 
 
-    public Product(int id, String name, int price, String owner) {
+    public Product(int id, String name, double price, String owner) {
         this.name = name;
         this.price = price;
         this.id = id;
@@ -62,11 +62,11 @@ public class Product {
         this.name = name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
